+++ src/star6e_runtime.c
@@ -552,6 +552,10 @@ static void start_custom_ae(const Star6ePipelineState *ps,
 	} else if (vcfg->isp.shutter_max_us > 0) {
 		ae_cfg.shutter_max_us = vcfg->isp.shutter_max_us;
 	}
+	if (vcfg->isp.gain_min > 0)
+		ae_cfg.gain_min = vcfg->isp.gain_min;
+	if (vcfg->isp.shutter_min_us > 0)
+		ae_cfg.shutter_min_us = vcfg->isp.shutter_min_us;
 	ae_cfg.verbose = vcfg->system.verbose;
 	star6e_cus3a_start(&ae_cfg);
 }